Moved components to components directory
