+++ src/app/app.component.ts
@@ -2,13 +2,13 @@ import { Component, OnInit, inject } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterOutlet } from '@angular/router';
 
-import { NavigationComponent } from './navigation/navigation.component';
-import { JumbotronComponent } from './jumbotron/jumbotron.component';
-import { ShortenUrlComponent } from './shorten-url/shorten-url.component';
-import { ShortenedLinksComponent } from './shortened-links/shortened-links.component';
-import { StatisticsComponent } from './statistics/statistics.component';
-import { GetStartedComponent } from './get-started/get-started.component';
-import { FooterComponent } from './footer/footer.component';
+import { NavigationComponent } from './components/navigation/navigation.component';
+import { JumbotronComponent } from './components/jumbotron/jumbotron.component';
+import { ShortenUrlComponent } from './components/shorten-url/shorten-url.component';
+import { ShortenedLinksComponent } from './components/shortened-links/shortened-links.component';
+import { StatisticsComponent } from './components/statistics/statistics.component';
+import { GetStartedComponent } from './components/get-started/get-started.component';
+import { FooterComponent } from './components/footer/footer.component';
 import { UrlService } from './services/url.service';
 import { LocalStorageService } from './services/local-storage.service';
 import { Url } from './models/url.model';
