+++ src/app/components/footer/footer.component.html
@@ -4,7 +4,7 @@
       <div class="col-lg-3 col-xl-3 p-0">
         <img
           class="img-fluid my-5 white-color"
-          src="../../assets/logo.svg"
+          src="../../../assets/logo.svg"
           alt="Shortly logo"
         />
       </div>
@@ -93,7 +93,7 @@
         >
           <img
             class="img-fluid"
-            src="../../assets/icon-facebook.svg"
+            src="../../../assets/icon-facebook.svg"
             alt="Facebook icon"
           />
         </a>
@@ -106,7 +106,7 @@
         >
           <img
             class="img-fluid"
-            src="../../assets/icon-twitter.svg"
+            src="../../../assets/icon-twitter.svg"
             alt="Twitter icon"
           />
         </a>
@@ -119,7 +119,7 @@
         >
           <img
             class="img-fluid"
-            src="../../assets/icon-pinterest.svg"
+            src="../../../assets/icon-pinterest.svg"
             alt="Pinterest icon"
           />
         </a>
@@ -132,7 +132,7 @@
         >
           <img
             class="img-fluid"
-            src="../../assets/icon-instagram.svg"
+            src="../../../assets/icon-instagram.svg"
             alt="Instagarm icon"
           />
         </a>
