+++ src/app/components/jumbotron/jumbotron.component.html
@@ -25,7 +25,7 @@
       >
         <img
           class="img-fluid image-heading"
-          src="../../assets/illustration-working.svg"
+          src="../../../assets/illustration-working.svg"
           alt="photo showing work"
         />
       </div>
