+++ src/app/components/jumbotron/jumbotron.component.scss
