+++ src/app/components/navigation/navigation.component.html
@@ -1,7 +1,7 @@
 <nav class="navbar navbar-expand-xl">
   <div class="container-fluid p-0">
     <a class="navbar-brand p-0 me-5" href="#"
-      ><img src="../../assets/logo.svg" alt="Shortly logo"
+      ><img src="../../../assets/logo.svg" alt="Shortly logo"
     /></a>
     <button
       class="navbar-toggler border-0 p-0"
