+++ src/app/components/navigation/navigation.component.scss
@@ -39,7 +39,7 @@ nav {
 }
 
 .navbar-toggler-icon {
-    background-image: url('../../assets/menu-sharp.svg');
+    background-image: url('../../../assets/menu-sharp.svg');
     filter: opacity(0.2);
 }
