+++ src/app/components/shortened-links/shortened-links.component.scss
