+++ src/app/components/statistics/statistics.component.html
@@ -22,7 +22,7 @@
           <div class="rounded-circle align-self-xl-start circle-background">
             <img
               class="img-fluid"
-              src="../../assets/icon-brand-recognition.svg"
+              src="../../../assets/icon-brand-recognition.svg"
               alt="Brand recognition icon"
             />
           </div>
@@ -46,7 +46,7 @@
           <div class="rounded-circle align-self-xl-start circle-background">
             <img
               class="img-fluid"
-              src="../../assets/icon-detailed-records.svg"
+              src="../../../assets/icon-detailed-records.svg"
               alt="Detailed records icon"
             />
           </div>
@@ -71,7 +71,7 @@
           <div class="rounded-circle align-self-xl-start circle-background">
             <img
               class="img-fluid"
-              src="../../assets/icon-fully-customizable.svg"
+              src="../../../assets/icon-fully-customizable.svg"
               alt="Fully customizable icon"
             />
           </div>
