+++ src/app/components/statistics/statistics.component.ts
